==> notebooks/03_sample_data_split.ipynb <==
## Setup enviroment

import os
import gc
import pandas as pd
import numpy as np
from catboost import CatBoostClassifier
from sklearn.metrics import log_loss, roc_auc_score, average_precision_score, brier_score_loss, precision_recall_curve
from sklearn.calibration import calibration_curve, CalibratedClassifierCV
from sklearn.inspection import permutation_importance
from sklearn.model_selection import train_test_split
import matplotlib.pyplot as plt
import seaborn as sns
import sys
sys.path.append(r'/home/jeanlr/projetos/lending-club/global')
from util import *
import arfs.feature_selection.allrelevant as arfsgroot
import shap
import json
import joblib

sns.set(style='whitegrid')

## Read the parquet file located at ../data/processed/feature_engineering_train.parquet

df = pd.read_parquet('../data/processed/feature_engineering_train.parquet')
df.info() 

## Split Out-Of-Time Train, Validation, Test

default_nao_nulo = df[df['default'].notnull()]
proporcao = default_nao_nulo['default'].value_counts(normalize=True)

# Exibir a proporção
print("Proporção de 0 e 1:")
print(proporcao)

## Remove nulls on target

# Remove as linhas com valores nulos na coluna 'default'
df = df.dropna(subset=['default'])

# Verifica a distribuição dos valores após a remoção
print(df['default'].value_counts(normalize=True))

train_max_date = '2016-01-01'
validation_max_date = '2016-10-01'
test_start_date = '2017-03-01'

# Loading features from features_list.json
with open('../artifacts/features_list.json', 'r') as f:
    features = json.load(f)

train_df = df.loc[df['issue_d'] < train_max_date, features]
validation_df = df.loc[(df['issue_d'] >= train_max_date) & (df['issue_d'] < validation_max_date), features]

# Get Train and Validation Subsamples to speed up the process
_, train_df = train_test_split(train_df, test_size=0.10, random_state=42, stratify=train_df['default'])

# Spliting the original validatio into two shuffled datasets one for validation only and the other for calibration
validation_df, calibration_df = train_test_split(validation_df, test_size=0.20, random_state=42, stratify=validation_df['default'])

print('Train Shape: ', train_df.shape)
print('Validation shape: ', validation_df.shape)
print('Calibration shape: ', calibration_df.shape)

train_df.columns

## Saving ABT's

# Save training data
train_df.to_parquet('../data/processed/train_df.parquet')

# Save validation data
validation_df.to_parquet('../data/processed/validation_df.parquet')

# Save calibration data
calibration_df.to_parquet('../data/processed/calibration_df.parquet')
