==> notebooks/08_scoring_test.ipynb <==
## Setup enviroment

import os
import gc
import pandas as pd
import numpy as np
from catboost import CatBoostClassifier
from sklearn.calibration import CalibratedClassifierCV
from sklearn.metrics import log_loss, roc_auc_score, average_precision_score, brier_score_loss
import joblib
from venn_abers import VennAbersCalibrator
import sys
sys.path.append(r'/home/jeanlr/projetos/lending-club/global')
from util import *
import json

## Read Test ABT

test_df = pd.read_parquet('../data/processed/abt_test.parquet')

selected_features = None
with open('../artifacts/selected_features_list.json', 'r') as f:
    selected_features = json.load(f)
print(selected_features)

X_test = test_df[selected_features]

X_test.shape

categorical_column_names = X_test.select_dtypes(include=['category']).columns
categorical_column_names

cat_features = ['grade', 'sub_grade', 'emp_length', 'home_ownership',
       'verification_status', 'application_type']

## Loading and aplly the calibrated_model_sigmoid

# Carregar o modelo salvo
calibrated_model_sigmoid = joblib.load("../artifacts/calibrated_model_sigmoid.pkl")

y_test_pred = calibrated_model_sigmoid.predict_proba(X_test)[:, 1]

y_test_pred

submission = test_df.reset_index()[["index"]].copy()  # Pegando o índice como coluna
submission.rename(columns={"index": "Id"}, inplace=True)  # Renomeando para "Id"

# Adicionando as previsões
submission["default"] = y_test_pred

submission

submission.to_parquet('../data/processed/jean_submission.parquet', index=False)
